requirements.txt
torch~=2.7.1
pillow~=11.2.1
torchvision~=0.22.1
scikit-learn~=1.7.0
matplotlib~=3.10.3
numpy~=2.3.1
networkx~=3.5
plotly~=6.2.0